Integration of Espruino ultrasonic data collection with Python code
The moving_avg_not_net is an adapted version of the existing neural net
program that just collects data and sends it to the Pi in the same way
that it would normally send data to the neural net.  The us_data_col.py
program collects that data and rotates the dog.